src/requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
tqdm
Pillow
seaborn
editdistance
datasets